add Verification & authorization to Tasks Controller

+++ src/TodoList.WebApi/Controllers/TaskController.cs
@@ -13,9 +13,12 @@ namespace ToDoLIstAPi.Controllers;
 public class TaskController : ControllerBase
 {
     private readonly ITaskService _taskService;
+    private readonly ILogger<TaskController> _logger;
 
-    public TaskController(ITaskService taskService)
+    public TaskController(ITaskService taskService , 
+        ILogger<TaskController> logger)
     {
+        _logger = logger;
         _taskService = taskService;
     }
 
@@ -28,7 +31,8 @@ public class TaskController : ControllerBase
         {
             return BadRequest("Id is invalid");
         }
-    
+        var adminName = HttpContext.User.Claims.FirstOrDefault(u => u.Type.Equals(ClaimTypes.Name))?.Value;
+        _logger.LogInformation($"admin {adminName} is requesting task with id {id}");
         var taskResponse = await _taskService.GetTaskAsync(id);
         return Ok(taskResponse);
     }
@@ -36,13 +40,16 @@ public class TaskController : ControllerBase
     [HttpGet("collection")]
     public async Task<IActionResult> GetTasks()
     {
+        var adminName = HttpContext.User.Claims.FirstOrDefault(u => u.Type.Equals(ClaimTypes.Name))?.Value;
+        _logger.LogInformation($"admin {adminName} is requesting all the tasks");
         var tasks = await _taskService.GetAllTaskAsync();
         return Ok(tasks);
     }
 
     [Authorize]
     [HttpPost]
-    public async Task<IActionResult> CreateTaskbyId([FromBody] TaskAddForUserDto inputTask)
+    // user create a task for himself
+    public async Task<IActionResult> CreateTaskForTheUser([FromBody] TaskAddForUserDto inputTask)
     {
         if (!ModelState.IsValid)
         {
@@ -51,19 +58,52 @@ public class TaskController : ControllerBase
         var userId = HttpContext.User.Claims.FirstOrDefault( u => u.Type.Equals(ClaimTypes.NameIdentifier))?.Value;
         if (userId is null)
             return BadRequest("User dosent Have an Id"); 
-        await _taskService.CreateTaskAsync(int.Parse(userId), inputTask);
+        inputTask.UserId = int.Parse(userId);
+        await _taskService.CreateTaskAsync(inputTask);
         return NoContent();
     }
-    [HttpPut]
-    public async Task<IActionResult> UpdateTaskbyId([FromBody] Tasks inputTask)
+    
+    [Authorize(Roles = "Admin")]
+    [HttpPost]
+    // admin create a task for a user
+    public async Task<IActionResult> CreateTaskByAdmin([FromBody] TaskAddForUserDto inputTask)
     {
         if (!ModelState.IsValid)
         {
             return UnprocessableEntity("Invalid input");
         }
+        var adminName = HttpContext.User.Claims.FirstOrDefault(u => u.Type.Equals(ClaimTypes.Name))?.Value;
+        _logger.LogInformation($"admin {adminName} is adding  a task for user with id {inputTask.UserId}");
+        await _taskService.CreateTaskAsync(inputTask);
+        return NoContent();
+    }
+    
+    
+    
+    [Authorize]
+    [HttpPut("{id}")]
+    // (Admin && user who is assigned to the task) can update the task
+    public async Task<IActionResult> UpdateTaskbyId([FromBody]  Tasks inputTask)
+    {
+        if (!ModelState.IsValid)
+        {
+            return UnprocessableEntity("Invalid input");
+        }
+        _logger.LogInformation("Checking if user is allowed to update the task");
+
+        var userId = HttpContext.User.Claims.FirstOrDefault( u => u.Type.Equals(ClaimTypes.NameIdentifier))?.Value;
+        if (userId is null)
+            return BadRequest("User dosent Have an Id");
+        var role = HttpContext.User.Claims.FirstOrDefault(u => u.Type.Equals(ClaimTypes.Role))?.Value;
+        if ( role != "Admin" && inputTask.UserId != int.Parse(userId))
+        {
+            return Unauthorized("You are not allowed to update this task");
+        }
+
         await _taskService.UpdateTaskAsync(inputTask);
         return NoContent();
     }
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteTaskbyId(int id)
     {
@@ -71,15 +111,46 @@ public class TaskController : ControllerBase
         {
             return BadRequest("Id is invalid");
         }
+        _logger.LogInformation("Checking if user is allowed to delete the task");
+        var task = await _taskService.GetTaskAsync(id);
+        
+        if ( task is null)
+        {
+            return BadRequest("Task not found");
+        }
+        var userId = HttpContext.User.Claims.FirstOrDefault( u => u.Type.Equals(ClaimTypes.NameIdentifier))?.Value;
+        if (userId is null)
+            return BadRequest("User dosent Have an Id");
+        var role = HttpContext.User.Claims.FirstOrDefault(u => u.Type.Equals(ClaimTypes.Role))?.Value;
+        if ( role != "Admin" && task.UserId != int.Parse(userId))
+        {
+            return Unauthorized("You are not allowed to delete this task");
+        }
         await _taskService.DeleteTaskAsync(id);
         return NoContent();
     }
+    [Authorize]
+
     [HttpPut("finish/{id}")]
     public async Task<IActionResult> FinishTaskbyId(int id)
     {
         if (id <= 0)
         {
             return BadRequest("Id is invalid");
+        }
+        _logger.LogInformation("Checking if user is allowed to finish the task");
+        var task = await _taskService.GetTaskAsync(id);
+        if ( task is null)
+        {
+            return BadRequest("Task not found");
+        }
+        var userId = HttpContext.User.Claims.FirstOrDefault( u => u.Type.Equals(ClaimTypes.NameIdentifier))?.Value;
+        if (userId is null)
+            return BadRequest("User dosent Have an Id");
+        var role = HttpContext.User.Claims.FirstOrDefault(u => u.Type.Equals(ClaimTypes.Role))?.Value;
+        if ( role != "Admin" && task.UserId != int.Parse(userId))
+        {
+            return Unauthorized("You are not allowed to finish this task");
         }
         await _taskService.FinishTaskAsync(id);
         return NoContent();
